Zero-datetime decoding
For mysql it is possible to use zero dates and datetimes if the
sql_mode is not set to strict.

https://dev.mysql.com/doc/refman/8.0/en/datetime.html

If we raise then the full query to sql will fail and the user will not
be able to retreive anything.

It should be up to the user to decide how to handle datetime columns
when not in strict mode.