Support `LOAD DATA LOCAL INFILE` operations
Note that there are security risks detailed here:
https://gwillem.gitlab.io/2019/01/20/sites-hacked-via-mysql-protocal-flaw/

The approach taken here is to require the user to explicitly
provide the data to send the server for this LOAD request. Note
that the filepath provided in the SQL LOAD statement (and
then requested by the server for the client to provide) is
intentionally ignored by this implementation.