Add Client.com_quit/1 and use it in ClientTest
This is to stop warnings like these:

    2020-11-05 15:54:54 128 [Warning] Aborted connection 128 to db: 'myxql_test' user: 'root' host: '172.18.0.1' (Got an error reading communication packets)

that happen on MariaDB.

Note, we have a separate `Client.disconnect/1` which simply closes the
socket. Perhaps in the future we should change `disconnect/1` to call
`com_quit/1` first. Finally, it would be also nice to find a way to
use `com_quit/1` when shutting down the pool to again stop polluting
server logs with these warnings.